requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-anymail==10.2
django-countries==7.5.1
django-timezone-field==5.1
ecdsa==0.18.0
frozenlist==1.4.0
glicko2==2.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ms-cv==0.1.1
multidict==6.0.4
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.4.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xbox-webapi==2.0.11
yarl==1.9.2
